--- v2/objective-c/YelpAPISample/OAuthConsumer/OAHMAC_SHA1SignatureProvider.m
@@ -1,58 +1,0 @@
-//
-//  OAHMAC_SHA1SignatureProvider.m
-//  OAuthConsumer
-//
-//  Created by Jon Crosby on 10/19/07.
-//  Copyright 2007 Kaboomerang LLC. All rights reserved.
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-
-
-#import "OAHMAC_SHA1SignatureProvider.h"
-
-#include "hmac.h"
-#include "Base64Transcoder.h"
-
-@implementation OAHMAC_SHA1SignatureProvider
-
-- (NSString *)name {
-    return @"HMAC-SHA1";
-}
-
-- (NSString *)signClearText:(NSString *)text withSecret:(NSString *)secret {
-    NSData *secretData = [[secret dataUsingEncoding:NSUTF8StringEncoding] retain];
-    NSData *clearTextData = [[text dataUsingEncoding:NSUTF8StringEncoding] retain];
-    unsigned char result[20];
-    hmac_sha1((unsigned char *)[clearTextData bytes], [clearTextData length], (unsigned char *)[secretData bytes], [secretData length], result);
-	[secretData release];
-	[clearTextData release];
-    
-    //Base64 Encoding
-    
-    char base64Result[32];
-    size_t theResultLength = 32;
-    Base64EncodeData(result, 20, base64Result, &theResultLength);
-    NSData *theData = [NSData dataWithBytes:base64Result length:theResultLength];
-    
-    NSString *base64EncodedResult = [[[NSString alloc] initWithData:theData encoding:NSUTF8StringEncoding] autorelease];
-    
-    return base64EncodedResult;
-}
-
-@end
