+++ v2/objective-c/YelpAPISample/YPAuth/Categories/NSString+URLEncoding.h
@@ -32,4 +32,5 @@
 - (NSString *)encodedURLParameterString;
 - (NSString *)decodedURLString;
 - (NSString *)removeQuotes;
+
 @end
