+++ v2/objective-c/YelpAPISample/YPAuth/Categories/NSString+URLEncoding.m
@@ -34,7 +34,7 @@
                                                                            NULL,                   // characters to leave unescaped (NULL = all escaped sequences are replaced)
                                                                            CFSTR("?=&+"),          // legal URL characters to be escaped (NULL = all legal characters are replaced)
                                                                            kCFStringEncodingUTF8); // encoding
-	return [result autorelease];
+  return result;
 }
 
 - (NSString *)encodedURLParameterString {
@@ -43,7 +43,7 @@
                                                                            NULL,
                                                                            CFSTR(":/=,!$&'()*+;[]@#?"),
                                                                            kCFStringEncodingUTF8);
-	return [result autorelease];
+	return result;
 }
 
 - (NSString *)decodedURLString {
@@ -51,23 +51,22 @@
 																						  (CFStringRef)self,
 																						  CFSTR(""),
 																						  kCFStringEncodingUTF8);
-	
-	return [result autorelease];
-	
+
+  return result;
 }
 
--(NSString *)removeQuotes
+- (NSString *)removeQuotes
 {
 	NSUInteger length = [self length];
-	NSString *ret = self;
+	NSString *string = self;
 	if ([self characterAtIndex:0] == '"') {
-		ret = [ret substringFromIndex:1];
+		string = [string substringFromIndex:1];
 	}
 	if ([self characterAtIndex:length - 1] == '"') {
-		ret = [ret substringToIndex:length - 2];
+		string = [string substringToIndex:length - 2];
 	}
-	
-	return ret;
+
+	return string;
 }
 
 @end
