+++ v2/objective-c/YelpAPISample/YPAuth/Crypto/Base64Transcoder.h
@@ -33,4 +33,3 @@ extern size_t EstimateBas64DecodedDataSize(size_t inDataSize);
 
 extern bool Base64EncodeData(const void *inInputData, size_t inInputDataSize, char *outOutputData, size_t *ioOutputDataSize);
 extern bool Base64DecodeData(const void *inInputData, size_t inInputDataSize, void *ioOutputData, size_t *ioOutputDataSize);
-
