+++ v2/objective-c/YelpAPISample/YPAuth/YPConsumer.h
@@ -1,0 +1,18 @@
+//
+//  YPConsumer.h
+//  YelpAPISample
+//
+//  Created by Thibaud Robelain on 16/10/14.
+//  Copyright (c) 2014 Yelp Inc. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface YPConsumer : NSObject
+
+@property (strong, nonatomic) NSString *key;
+@property (strong, nonatomic) NSString *secret;
+
+- (instancetype)initWithKey:(NSString *)key secret:(NSString *)secret;
+
+@end
