+++ v2/objective-c/YelpAPISample/YPAuth/YPHMAC_SHA1SignatureProvider.h
@@ -1,0 +1,14 @@
+//
+//  YPHMAC_SHA1SignatureProvider.h
+//  YelpAPISample
+//
+//  Created by Thibaud Robelain on 16/10/14.
+//  Copyright (c) 2014 Yelp Inc. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import "YPSignatureProvider.h"
+
+@interface YPHMAC_SHA1SignatureProvider : NSObject<YPSignatureProvider>
+
+@end
